fix(browser): bind a client-hosted page's guest when the page is created
The browser manager learned a client page's guest only when an agent automation
command arrived, so for a user-driven page popup downloads resolved no owning
page and failed closed, and no download event reached the renderer at all.
Page creation now binds the guest and retirement releases it.

The download suite registered the guest itself, which hid all of that; it now
drives the same binding production uses. Executor tests split by topic to stay
under the line ceiling.
